Split out GAE deps from standard utils; client tools should not require the App Engine SDK.
