+++ rest-api/api/metrics_fields_api.py
@@ -1,4 +1,4 @@
-import api_util
+import app_util
 import offline.metrics_config
 
 from api_util import HEALTHPRO
@@ -7,6 +7,6 @@ from flask.ext.restful import Resource
 class MetricsFieldsApi(Resource):
   """API that returns the names and valid values for metric fields."""
 
-  @api_util.auth_required(HEALTHPRO)
+  @app_util.auth_required(HEALTHPRO)
   def get(self):
     return offline.metrics_config.get_fields()
