+++ rest-api/api/participant_api.py
@@ -1,4 +1,4 @@
-import api_util
+import app_util
 
 from api.base_api import UpdatableApi
 from api_util import PTC, PTC_AND_HEALTHPRO
@@ -8,14 +8,14 @@ class ParticipantApi(UpdatableApi):
   def __init__(self):
     super(ParticipantApi, self).__init__(ParticipantDao())
 
-  @api_util.auth_required(PTC_AND_HEALTHPRO)
+  @app_util.auth_required(PTC_AND_HEALTHPRO)
   def get(self, p_id):
     return super(ParticipantApi, self).get(p_id)
 
-  @api_util.auth_required(PTC)
+  @app_util.auth_required(PTC)
   def post(self):
     return super(ParticipantApi, self).post()
 
-  @api_util.auth_required(PTC)
+  @app_util.auth_required(PTC)
   def put(self, p_id):
     return super(ParticipantApi, self).put(p_id)
