+++ rest-api/api/questionnaire_api.py
@@ -1,4 +1,4 @@
-import api_util
+import app_util
 
 from api.base_api import UpdatableApi
 from api_util import PTC
@@ -12,7 +12,7 @@ class QuestionnaireApi(UpdatableApi):
   def __init__(self):
     super(QuestionnaireApi, self).__init__(QuestionnaireDao())
 
-  @api_util.auth_required(PTC)
+  @app_util.auth_required(PTC)
   def get(self, id_=None):
     if id_:
       return super(QuestionnaireApi, self).get(id_)
@@ -28,11 +28,10 @@ class QuestionnaireApi(UpdatableApi):
         raise NotFound('Could not find questionnaire with concept: %s' % concept)
       return self._make_response(questionnaire)
 
-  @api_util.auth_required(PTC)
+  @app_util.auth_required(PTC)
   def post(self):
     return super(QuestionnaireApi, self).post()
 
-  @api_util.auth_required(PTC)
+  @app_util.auth_required(PTC)
   def put(self, id_):
     return super(QuestionnaireApi, self).put(id_)
-
