+++ rest-api/api/questionnaire_response_api.py
@@ -1,4 +1,4 @@
-import api_util
+import app_util
 
 from api.base_api import BaseApi
 from api_util import PTC
@@ -8,11 +8,11 @@ class QuestionnaireResponseApi(BaseApi):
   def __init__(self):
     super(QuestionnaireResponseApi, self).__init__(QuestionnaireResponseDao())
 
-  @api_util.auth_required(PTC)
+  @app_util.auth_required(PTC)
   def get(self, p_id, id_):
     #pylint: disable=unused-argument
     return super(QuestionnaireResponseApi, self).get(id_)
 
-  @api_util.auth_required(PTC)
+  @app_util.auth_required(PTC)
   def post(self, p_id):
     return super(QuestionnaireResponseApi, self).post(participant_id=p_id)
