+++ rest-api/model/utils.py
@@ -1,4 +1,3 @@
-import config
 from dateutil.tz import tzutc
 from query import PropertyType
 from sqlalchemy.types import SmallInteger, TypeDecorator, DateTime
@@ -51,20 +50,6 @@ def from_client_participant_id(participant_id):
     return int(participant_id[1:])
   except ValueError:
     raise BadRequest("Invalid participant ID: %s" % participant_id)
-
-def get_biobank_id_prefix():
-  return config.getSetting(config.BIOBANK_ID_PREFIX)
-
-def to_client_biobank_id(biobank_id):
-  return '%s%d' % (get_biobank_id_prefix(), biobank_id)
-
-def from_client_biobank_id(biobank_id):
-  if not biobank_id.startswith(get_biobank_id_prefix()):
-    raise BadRequest("Invalid biobank ID: %s" % biobank_id)
-  try:
-    return int(biobank_id[1:])
-  except ValueError:
-    raise BadRequest("Invalid biobank ID: %s" % biobank_id)
 
 class ParticipantIdConverter(BaseConverter):
