+++ rest-api/test/test_data.py
@@ -9,7 +9,8 @@ import random
 from code_constants import PPI_SYSTEM, CONSENT_FOR_STUDY_ENROLLMENT_MODULE, BIOBANK_TESTS
 from code_constants import FIRST_NAME_QUESTION_CODE, LAST_NAME_QUESTION_CODE, EMAIL_QUESTION_CODE
 from model.code import Code, CodeType
-from model.utils import to_client_participant_id, to_client_biobank_id
+from model.config_utils import to_client_biobank_id
+from model.utils import to_client_participant_id
 
 def consent_code():
   return Code(system=PPI_SYSTEM, value=CONSENT_FOR_STUDY_ENROLLMENT_MODULE,
